Fix starting GO term filtering
Before, only GO terms with zero children were retrieved from a list of all GO terms associated with the genes of interest, but this would omit terms with children if none of their children was directly associated with the input genes. Now the term list is reduced by internally checking if any term in the list is the parent of another term, and removing the parent term.